// interfaces/igdbResponse.ts
export default interface igdbResponse 
{
    id:                      number;
    age_ratings:             number[];
    aggregated_rating:       number;
    aggregated_rating_count: number;
    alternative_names:       number[];
    artworks:                number[];
    category:                number;
    collection:              number;
    cover:                   number;
    created_at:              number;
    external_games:          number[];
    first_release_date:      number;
    franchises:              number[];
    game_modes:              number[];
    genres:                  number[];
    involved_companies:      number[];
    keywords:                number[];
    multiplayer_modes:       number[];
    name:                    string;
    platforms:               number[];
    player_perspectives:     number[];
    rating:                  number;
    rating_count:            number;
    release_dates:           number[];
    screenshots:             number[];
    similar_games:           number[];
    slug:                    string;
    summary:                 string;
    tags:                    number[];
    themes:                  number[];
    total_rating:            number;
    total_rating_count:      number;
    updated_at:              number;
    url:                     string;
    version_parent:          number;
    version_title:           string;
    websites:                number[];
    checksum:                string;
    language_supports:       number[];
    game_localizations:      number[];
    collections:             number[];
}